+++ Log.log
@@ -1,0 +1,18 @@
+2020/09/30 10:46:38 config.go:87: REPLACE:503 Too busy, try again laterリトライ
+2020/09/30 10:46:38 config.go:87: REPLACE:503 Too busy, try again laterリトライ
+2020/09/30 10:46:38 config.go:87: REPLACE:503 Too busy, try again laterリトライ
+2020/09/30 10:46:39 main.go:57: mkdir REPLACE\202009: The system cannot find the path specified.
+2020/09/30 10:56:34 main.go:42: yaml: line 10: did not find expected '-' indicator
+2020/09/30 10:57:11 main.go:57: open D:\202009\サーバールーム\_\20200927\_\20200927\.csv: The system cannot find the path specified.
+2020/09/30 15:53:34 main.go:57: open D:\202009\サーバールーム_\20200927_\20200927.csv: The system cannot find the path specified.
+2020/09/30 15:55:28 main.go:57: open D:\202009\サーバールーム_20200927_20200927.csv: The system cannot find the path specified.
+2020/09/30 15:57:43 main.go:57: open D:\202009\サーバールーム_20200927_20200927.csv: The system cannot find the path specified.
+2020/09/30 15:59:27 main.go:57: open D:\202009\サーバールーム_20200927_20200927.csv: The system cannot find the path specified.
+2020/09/30 15:59:31 main.go:57: open D:\202009\サーバールーム_20200927_20200927.csv: The system cannot find the path specified.
+2020/09/30 16:01:46 main.go:57: mkdir D\202009: The system cannot find the path specified.
+2020/09/30 16:04:35 main.go:57: mkdir D\202009: The system cannot find the path specified.
+2020/09/30 16:05:18 config.go:87: サーバールーム:503 Too busy, try again laterリトライ
+2020/09/30 16:05:18 config.go:87: サーバールーム:503 Too busy, try again laterリトライ
+2020/09/30 16:05:18 config.go:87: サーバールーム:503 Too busy, try again laterリトライ
+2020/09/30 16:05:18 config.go:87: サーバールーム:503 Too busy, try again laterリトライ
+2020/09/30 16:05:18 config.go:87: サーバールーム:503 Too busy, try again laterリトライ
