+++ planexdownload.code-workspace
@@ -1,0 +1,7 @@
+{
+    "folders": [
+        {
+            "path": "."
+        }
+    ]
+}
